requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
tqdm==4.66.2
python-dotenv==1.0.1
sentence-transformers==2.7.0
faiss-cpu==1.8.0
numpy==1.26.4
openai==1.14.2
pyyaml==6.0.1
scikit-learn==1.4.2
fastapi==0.110.0
uvicorn[standard]==0.29.0
pdfplumber==0.11.0
boto3==1.34.57
notion-client==2.2.0

pymilvus==2.4.1
httpx==0.27.0
python-docx==1.1.0
python-pptx==0.6.23
pillow==10.2.0
pytesseract==0.3.10
pytest
requests-mock